refactor(CartItem): 코드 reformat
