+++ src/main/java/gift/domain/cartItem/CartItemService.java
@@ -1,13 +1,11 @@
-package gift.domain.cart;
+package gift.domain.cartItem;
 
 import gift.domain.product.JpaProductRepository;
 import gift.domain.product.Product;
 import gift.domain.user.JpaUserRepository;
 import gift.domain.user.User;
-
 import gift.global.exception.BusinessException;
 import java.util.List;
-import java.util.Optional;
 import java.util.stream.Collectors;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
@@ -46,7 +44,8 @@ public class CartItemService {
 
         // 기존에 존재하면 update
         if (cartItemRepository.existsByUserAndProduct(user, product)) {
-            CartItem existsCartItem = cartItemRepository.findByUserIdAndProductId(userId, productId).get();
+            CartItem existsCartItem = cartItemRepository.findByUserIdAndProductId(userId, productId)
+                .get();
             existsCartItem.addOneMore();
             return existsCartItem.getCount();
         }
@@ -92,7 +91,8 @@ public class CartItemService {
      */
     @Transactional
     public int updateCartItem(Long userId, Long productId, int count) {
-        CartItem findCartItem = cartItemRepository.findByUserIdAndProductId(userId, productId).orElseThrow(() -> new BusinessException(HttpStatus.BAD_REQUEST, "해당 상품이 존재하지 않습니다."));
+        CartItem findCartItem = cartItemRepository.findByUserIdAndProductId(userId, productId)
+            .orElseThrow(() -> new BusinessException(HttpStatus.BAD_REQUEST, "해당 상품이 존재하지 않습니다."));
         findCartItem.updateCount(count); // 수량 수정
         return count;
     }
